Move tests to propagator
