+++ ethcore/sync/src/chain/mod.rs
@@ -1194,7 +1194,7 @@ pub mod tests {
 		sync
 	}
 
-	fn insert_dummy_peer(sync: &mut ChainSync, peer_id: PeerId, peer_latest_hash: H256) {
+	pub fn insert_dummy_peer(sync: &mut ChainSync, peer_id: PeerId, peer_latest_hash: H256) {
 		sync.peers.insert(peer_id,
 			PeerInfo {
 				protocol_version: 0,
@@ -1245,300 +1245,6 @@ pub mod tests {
 		// size of three rlp encoded hash-difficulty
 		assert_eq!(107, rlp.len());
 	}
-
-	#[test]
-	fn sends_new_hashes_to_lagging_peer() {
-		let mut client = TestBlockChainClient::new();
-		client.add_blocks(100, EachBlockWith::Uncle);
-		let queue = RwLock::new(VecDeque::new());
-		let mut sync = dummy_sync_with_peer(client.block_hash_delta_minus(5), &client);
-		let chain_info = client.chain_info();
-		let ss = TestSnapshotService::new();
-		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-
-		let peers = sync.get_lagging_peers(&chain_info);
-		let peer_count = SyncPropagator::propagate_new_hashes(&mut sync, &chain_info, &mut io, &peers);
-
-		// 1 message should be send
-		assert_eq!(1, io.packets.len());
-		// 1 peer should be updated
-		assert_eq!(1, peer_count);
-		// NEW_BLOCK_HASHES_PACKET
-		assert_eq!(0x01, io.packets[0].packet_id);
-	}
-
-	#[test]
-	fn sends_latest_block_to_lagging_peer() {
-		let mut client = TestBlockChainClient::new();
-		client.add_blocks(100, EachBlockWith::Uncle);
-		let queue = RwLock::new(VecDeque::new());
-		let mut sync = dummy_sync_with_peer(client.block_hash_delta_minus(5), &client);
-		let chain_info = client.chain_info();
-		let ss = TestSnapshotService::new();
-		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		let peers = sync.get_lagging_peers(&chain_info);
-		let peer_count = SyncPropagator::propagate_blocks(&mut sync, &chain_info, &mut io, &[], &peers);
-
-		// 1 message should be send
-		assert_eq!(1, io.packets.len());
-		// 1 peer should be updated
-		assert_eq!(1, peer_count);
-		// NEW_BLOCK_PACKET
-		assert_eq!(0x07, io.packets[0].packet_id);
-	}
-
-	#[test]
-	fn sends_sealed_block() {
-		let mut client = TestBlockChainClient::new();
-		client.add_blocks(100, EachBlockWith::Uncle);
-		let queue = RwLock::new(VecDeque::new());
-		let hash = client.block_hash(BlockId::Number(99)).unwrap();
-		let mut sync = dummy_sync_with_peer(client.block_hash_delta_minus(5), &client);
-		let chain_info = client.chain_info();
-		let ss = TestSnapshotService::new();
-		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		let peers = sync.get_lagging_peers(&chain_info);
-		let peer_count = SyncPropagator::propagate_blocks(&mut sync ,&chain_info, &mut io, &[hash.clone()], &peers);
-
-		// 1 message should be send
-		assert_eq!(1, io.packets.len());
-		// 1 peer should be updated
-		assert_eq!(1, peer_count);
-		// NEW_BLOCK_PACKET
-		assert_eq!(0x07, io.packets[0].packet_id);
-	}
-
-	#[test]
-	fn sends_proposed_block() {
-		let mut client = TestBlockChainClient::new();
-		client.add_blocks(2, EachBlockWith::Uncle);
-		let queue = RwLock::new(VecDeque::new());
-		let block = client.block(BlockId::Latest).unwrap().into_inner();
-		let mut sync = ChainSync::new(SyncConfig::default(), &client, Arc::new(NoopPrivateTxHandler));
-		sync.peers.insert(0,
-			PeerInfo {
-				// Messaging protocol
-				protocol_version: 2,
-				genesis: H256::zero(),
-				network_id: 0,
-				latest_hash: client.block_hash_delta_minus(1),
-				difficulty: None,
-				asking: PeerAsking::Nothing,
-				asking_blocks: Vec::new(),
-				asking_hash: None,
-				ask_time: Instant::now(),
-				last_sent_transactions: HashSet::new(),
-				expired: false,
-				confirmation: super::ForkConfirmation::Confirmed,
-				snapshot_number: None,
-				snapshot_hash: None,
-				asking_snapshot_data: None,
-				block_set: None,
-			});
-		let ss = TestSnapshotService::new();
-		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		SyncPropagator::propagate_proposed_blocks(&mut sync, &mut io, &[block]);
-
-		// 1 message should be sent
-		assert_eq!(1, io.packets.len());
-		// NEW_BLOCK_PACKET
-		assert_eq!(0x07, io.packets[0].packet_id);
-	}
-
-	#[test]
-	fn propagates_transactions() {
-		let mut client = TestBlockChainClient::new();
-		client.add_blocks(100, EachBlockWith::Uncle);
-		client.insert_transaction_to_queue();
-		let mut sync = dummy_sync_with_peer(client.block_hash_delta_minus(1), &client);
-		let queue = RwLock::new(VecDeque::new());
-		let ss = TestSnapshotService::new();
-		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		let peer_count = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-		// Try to propagate same transactions for the second time
-		let peer_count2 = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-		// Even after new block transactions should not be propagated twice
-		sync.chain_new_blocks(&mut io, &[], &[], &[], &[], &[], &[]);
-		// Try to propagate same transactions for the third time
-		let peer_count3 = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-
-		// 1 message should be send
-		assert_eq!(1, io.packets.len());
-		// 1 peer should be updated but only once
-		assert_eq!(1, peer_count);
-		assert_eq!(0, peer_count2);
-		assert_eq!(0, peer_count3);
-		// TRANSACTIONS_PACKET
-		assert_eq!(0x02, io.packets[0].packet_id);
-	}
-
-	#[test]
-	fn does_not_propagate_new_transactions_after_new_block() {
-		let mut client = TestBlockChainClient::new();
-		client.add_blocks(100, EachBlockWith::Uncle);
-		client.insert_transaction_to_queue();
-		let mut sync = dummy_sync_with_peer(client.block_hash_delta_minus(1), &client);
-		let queue = RwLock::new(VecDeque::new());
-		let ss = TestSnapshotService::new();
-		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		let peer_count = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-		io.chain.insert_transaction_to_queue();
-		// New block import should not trigger propagation.
-		// (we only propagate on timeout)
-		sync.chain_new_blocks(&mut io, &[], &[], &[], &[], &[], &[]);
-
-		// 2 message should be send
-		assert_eq!(1, io.packets.len());
-		// 1 peer should receive the message
-		assert_eq!(1, peer_count);
-		// TRANSACTIONS_PACKET
-		assert_eq!(0x02, io.packets[0].packet_id);
-	}
-
-	#[test]
-	fn does_not_fail_for_no_peers() {
-		let mut client = TestBlockChainClient::new();
-		client.add_blocks(100, EachBlockWith::Uncle);
-		client.insert_transaction_to_queue();
-		// Sync with no peers
-		let mut sync = ChainSync::new(SyncConfig::default(), &client, Arc::new(NoopPrivateTxHandler));
-		let queue = RwLock::new(VecDeque::new());
-		let ss = TestSnapshotService::new();
-		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		let peer_count = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-		sync.chain_new_blocks(&mut io, &[], &[], &[], &[], &[], &[]);
-		// Try to propagate same transactions for the second time
-		let peer_count2 = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-
-		assert_eq!(0, io.packets.len());
-		assert_eq!(0, peer_count);
-		assert_eq!(0, peer_count2);
-	}
-
-	#[test]
-	fn propagates_transactions_without_alternating() {
-		let mut client = TestBlockChainClient::new();
-		client.add_blocks(100, EachBlockWith::Uncle);
-		client.insert_transaction_to_queue();
-		let mut sync = dummy_sync_with_peer(client.block_hash_delta_minus(1), &client);
-		let queue = RwLock::new(VecDeque::new());
-		let ss = TestSnapshotService::new();
-		// should sent some
-		{
-			let mut io = TestIo::new(&mut client, &ss, &queue, None);
-			let peer_count = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-			assert_eq!(1, io.packets.len());
-			assert_eq!(1, peer_count);
-		}
-		// Insert some more
-		client.insert_transaction_to_queue();
-		let (peer_count2, peer_count3) = {
-			let mut io = TestIo::new(&mut client, &ss, &queue, None);
-			// Propagate new transactions
-			let peer_count2 = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-			// And now the peer should have all transactions
-			let peer_count3 = SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-			(peer_count2, peer_count3)
-		};
-
-		// 2 message should be send (in total)
-		assert_eq!(2, queue.read().len());
-		// 1 peer should be updated but only once after inserting new transaction
-		assert_eq!(1, peer_count2);
-		assert_eq!(0, peer_count3);
-		// TRANSACTIONS_PACKET
-		assert_eq!(0x02, queue.read()[0].packet_id);
-		assert_eq!(0x02, queue.read()[1].packet_id);
-	}
-
-	#[test]
-	fn should_maintain_transations_propagation_stats() {
-		let mut client = TestBlockChainClient::new();
-		client.add_blocks(100, EachBlockWith::Uncle);
-		client.insert_transaction_to_queue();
-		let mut sync = dummy_sync_with_peer(client.block_hash_delta_minus(1), &client);
-		let queue = RwLock::new(VecDeque::new());
-		let ss = TestSnapshotService::new();
-		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-		SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-
-		let stats = sync.transactions_stats();
-		assert_eq!(stats.len(), 1, "Should maintain stats for single transaction.")
-	}
-
-	#[test]
-	fn should_propagate_service_transaction_to_selected_peers_only() {
-		let mut client = TestBlockChainClient::new();
-		client.insert_transaction_with_gas_price_to_queue(U256::zero());
-		let block_hash = client.block_hash_delta_minus(1);
-		let mut sync = ChainSync::new(SyncConfig::default(), &client, Arc::new(NoopPrivateTxHandler));
-		let queue = RwLock::new(VecDeque::new());
-		let ss = TestSnapshotService::new();
-		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-
-		// when peer#1 is Geth
-		insert_dummy_peer(&mut sync, 1, block_hash);
-		io.peers_info.insert(1, "Geth".to_owned());
-		// and peer#2 is Parity, accepting service transactions
-		insert_dummy_peer(&mut sync, 2, block_hash);
-		io.peers_info.insert(2, "Parity/v1.6".to_owned());
-		// and peer#3 is Parity, discarding service transactions
-		insert_dummy_peer(&mut sync, 3, block_hash);
-		io.peers_info.insert(3, "Parity/v1.5".to_owned());
-		// and peer#4 is Parity, accepting service transactions
-		insert_dummy_peer(&mut sync, 4, block_hash);
-		io.peers_info.insert(4, "Parity/v1.7.3-ABCDEFGH".to_owned());
-
-		// and new service transaction is propagated to peers
-		SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-
-		// peer#2 && peer#4 are receiving service transaction
-		assert!(io.packets.iter().any(|p| p.packet_id == 0x02 && p.recipient == 2)); // TRANSACTIONS_PACKET
-		assert!(io.packets.iter().any(|p| p.packet_id == 0x02 && p.recipient == 4)); // TRANSACTIONS_PACKET
-		assert_eq!(io.packets.len(), 2);
-	}
-
-	#[test]
-	fn should_propagate_service_transaction_is_sent_as_separate_message() {
-		let mut client = TestBlockChainClient::new();
-		let tx1_hash = client.insert_transaction_to_queue();
-		let tx2_hash = client.insert_transaction_with_gas_price_to_queue(U256::zero());
-		let block_hash = client.block_hash_delta_minus(1);
-		let mut sync = ChainSync::new(SyncConfig::default(), &client, Arc::new(NoopPrivateTxHandler));
-		let queue = RwLock::new(VecDeque::new());
-		let ss = TestSnapshotService::new();
-		let mut io = TestIo::new(&mut client, &ss, &queue, None);
-
-		// when peer#1 is Parity, accepting service transactions
-		insert_dummy_peer(&mut sync, 1, block_hash);
-		io.peers_info.insert(1, "Parity/v1.6".to_owned());
-
-		// and service + non-service transactions are propagated to peers
-		SyncPropagator::propagate_new_transactions(&mut sync, &mut io);
-
-		// two separate packets for peer are queued:
-		// 1) with non-service-transaction
-		// 2) with service transaction
-		let sent_transactions: Vec<UnverifiedTransaction> = io.packets.iter()
-			.filter_map(|p| {
-				if p.packet_id != 0x02 || p.recipient != 1 { // TRANSACTIONS_PACKET
-					return None;
-				}
-
-				let rlp = Rlp::new(&*p.data);
-				let item_count = rlp.item_count().unwrap_or(0);
-				if item_count != 1 {
-					return None;
-				}
-
-				rlp.at(0).ok().and_then(|r| r.as_val().ok())
-			})
-			.collect();
-		assert_eq!(sent_transactions.len(), 2);
-		assert!(sent_transactions.iter().any(|tx| tx.hash() == tx1_hash));
-		assert!(sent_transactions.iter().any(|tx| tx.hash() == tx2_hash));
-	}
-
 	// idea is that what we produce when propagading latest hashes should be accepted in
 	// on_peer_new_hashes in our code as well
 	#[test]
